+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [1-alpha2] - 2023-02-21
+- python matplotlib support
+
 ## [1-alpha1] - 2023-02-16
 - initial version (with a simple tech.ml.dataset-Hanami adapter)